+++ example/helper/ClassHelperExample.php
@@ -2,11 +2,18 @@
 
 require __DIR__.'/../../vendor/autoload.php';
 
+use LDL\Framework\Base\Collection\Contracts\AppendableInterface;
+use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
+use LDL\Framework\Base\Collection\Traits\AppendableInterfaceTrait;
+use LDL\Framework\Base\Collection\Traits\AppendManyTrait;
+use LDL\Framework\Base\Collection\Traits\CollectionInterfaceTrait;
 use LDL\Framework\Helper\ClassHelper;
 
-class Test
+class Test implements CollectionInterface, AppendableInterface
 {
-    use \LDL\Framework\Base\Collection\Traits\AppendableInterfaceTrait;
+    use CollectionInterfaceTrait;
+    use AppendableInterfaceTrait;
+    use AppendManyTrait;
 }
 
 echo "Get CallableCollection traits\n";